Step 0: paint startup from committed projections, off-thread probes
Startup no longer walks the archive or the playable root when a
committed projection exists: the archive refresh is gated on
archive_profiles.indexed_at (new db helper archive_profile_indexed_at),
and the folder scan runs only when the RootOpened reply shows an
unprojected root. Migration probes move off the UI constructor into the
startup thread (new StartupStatus message shows the modal only when a
real migration is needed), the legacy JSON cache probe no longer opens
a database connection first, and coarse timing logs cover each stage.
Measured on the network-mounted library: background preparation drops
from ~15 s of scanning to ~55 ms.

Also: fix retro-junk-io statvfs build breakage on macOS (u32 fields),
give collection rows one whole-row response instead of a per-cell
union (fixes synthetic-click accessibility rect), and repair two
tests (tempdir canonicalization; bounded harness steps).

+++ retro-junk-io/src/lib.rs
@@ -339,7 +339,13 @@ fn available_space(path: &Path) -> Result<Option<u64>, StageError> {
     }
     // SAFETY: a successful `statvfs` call initialized the structure.
     let stats = unsafe { stats.assume_init() };
-    Ok(Some(stats.f_bavail.saturating_mul(stats.f_frsize)))
+    // `statvfs` field widths differ per platform (u32 on macOS, u64 on
+    // Linux); `u64::from` widens or is the identity accordingly, so the
+    // conversion is only "useless" on one platform at a time.
+    #[allow(clippy::useless_conversion)]
+    Ok(Some(
+        u64::from(stats.f_bavail).saturating_mul(u64::from(stats.f_frsize)),
+    ))
 }
 
 #[cfg(not(unix))]
